MISC: [round buffer] RoundBuffer moved to utils folder
